fix cyclic import, put density compensation in separate file
