+++ mri/operators/fourier/utils/__init__.py
@@ -9,3 +9,4 @@
 
 from .orc_coefficients import *
 from .processing import *
+from .density import *
